Setup docs route to return general route info, removed controller.js(placeholder)
