--- Server/controllers/controller.js
@@ -1,1 +1,0 @@
-// This is just a placeholder for file structure
