+++ Server/routes/docs.route.js
@@ -1,7 +1,11 @@
 const express = require('express');
 const docsRouter = express.Router();
-const DocsController = require('../controllers/docs.controller');
+const DocsApiController = require('../controllers/docs.api.controller');
 
-docsRouter.route('/api').get(DocsController.listAllRoutes);
+docsRouter.route('/api').get(DocsApiController.listAllRoutes); //Returns an array with all the possible routes
+
+docsRouter
+    .route('/api/:owned?/:user?/:api?/:ingredient?') //The ? makes the param optional. Must have at least one parameter.
+    .get(DocsApiController.detailedRoute);
 
 module.exports = docsRouter;
